worker-service/requirements.txt
aio-pika
sqlalchemy
asyncpg
Pillow
aiohttp
psycopg2-binary
pytest
pytest-asyncio
